Use floating point numbers instead of ints

+++ app/controllers/recipe/RecipeCreation.scala
@@ -7,7 +7,7 @@ import io.scalaland.chimney.Transformer
 case class RecipeCreation(
     name: String,
     description: Option[String],
-    numberOfServings: Int
+    numberOfServings: BigDecimal
 )
 
 object RecipeCreation {
