+++ app/services/recipe/RecipeCreation.scala
@@ -5,7 +5,7 @@ import services.RecipeId
 case class RecipeCreation(
     name: String,
     description: Option[String],
-    numberOfServings: Int
+    numberOfServings: BigDecimal
 )
 
 object RecipeCreation {
